Put the slice-manipulation func in a separate file
